should allow for more than 256x256 map

+++ day-8.rs
@@ -4,7 +4,7 @@ use std::collections::HashSet;
 fn main() {
     let file_str = fs::read_to_string("day-8.input").expect("Failed to read file");
 
-    let mut map : Vec<Vec<u8>> = Vec::new();
+    let mut map : Vec<Vec<u32>> = Vec::new();
 
     for line in file_str.trim().split('\n') {
         let mut row = Vec::new();
@@ -14,16 +14,16 @@ fn main() {
         map.push(row);
     }
 
-    let mut wins : HashSet<(u8, u8)> = HashSet::new();
+    let mut wins : HashSet<(u32, u32)> = HashSet::new();
 
     let map_width = map[0].len();
 
     map.iter().enumerate().fold(
         (0..map_width).map(|_| b'0' - 1).collect(),
-        |maxes : Vec<u8>, (x, row)| {
+        |maxes : Vec<u32>, (x, row)| {
             row.iter().zip(maxes).enumerate().map(|(y, (&height, prev_max))| {
                 if height > prev_max {
-                    wins.insert((x as u8, y as u8));
+                    wins.insert((x as u32, y as u32));
                     return height;
                 }
                 return prev_max;
@@ -33,10 +33,10 @@ fn main() {
 
     map.iter().enumerate().rev().fold(
         (0..map_width).map(|_| b'0' - 1).collect(),
-        |maxes : Vec<u8>, (x, row)| {
+        |maxes : Vec<u32>, (x, row)| {
             row.iter().zip(maxes).enumerate().map(|(y, (&height, prev_max))| {
                 if height > prev_max {
-                    wins.insert((x as u8, y as u8));
+                    wins.insert((x as u32, y as u32));
                     return height;
                 }
                 return prev_max;
@@ -48,7 +48,7 @@ fn main() {
         |(x, row)| {
             row.iter().enumerate().fold(b'0' - 1, |prev_max, (y, &height)| {
                 if height > prev_max {
-                    wins.insert((x as u8, y as u8));
+                    wins.insert((x as u32, y as u32));
                     return height;
                 }
                 return prev_max;
@@ -60,7 +60,7 @@ fn main() {
         |(x, row)| {
             row.iter().enumerate().rev().fold(b'0' - 1, |prev_max, (y, &height)| {
                 if height > prev_max {
-                    wins.insert((x as u8, y as u8));
+                    wins.insert((x as u32, y as u32));
                     return height;
                 }
                 return prev_max;
